Default port to 2022, updating documentation to match new default. Addresses nice to have in #76.
Also print bind information on startup. An alternative to this approach would be to use info level logging, but the issue makes me think we want it all the time. I altered the README to show use of non-default port 22 which makes the note about root/sudo below still make sense.